requirements.txt
aiokafka==0.10.0
asyncio==3.4.3
faust==1.10.4
faust-streaming==0.11.0
Flask==3.0.3
inference==0.12.0
kafka-python==1.4.7
numpy==1.26.4
opencv-python==4.8.0.76
robinhood-aiokafka==1.1.6
roboflow==1.1.32
supervision==0.21.0
